refactor: cpp and python unit tests were refactored
