+++ tests/cli/check_subcommand/cpp.rs
@@ -1,0 +1,47 @@
+/*
+*  Quick Test: CLI for stress testing in competitive programming
+*  Copyright (C) 2021-present / Luis Miguel Báez
+*  License: MIT (See the LICENSE file in the repository root directory)
+*/
+
+use std::{error::Error, process::Command};
+
+use assert_cmd::assert::OutputAssertExt;
+use predicates::prelude::predicate;
+
+use crate::util::{
+    test_command_handler::execute_command_check,
+    test_constants::{BINARY, CHECKER_FILE_CPP, FOLDER_CHECK, GEN_FILE_CPP, TARGET_FILE_CPP},
+    test_utilities::create_files_check,
+};
+
+use super::codes::{CHECKER_CPP_CHECK, GEN_CPP_CHECK, TARGET_CPP_CHECK};
+
+#[test]
+fn cmd_check_target_cpp_check_cpp_gen_cpp() -> Result<(), Box<dyn Error>> {
+    create_files_check(
+        TARGET_FILE_CPP,
+        CHECKER_FILE_CPP,
+        GEN_FILE_CPP,
+        TARGET_CPP_CHECK,
+        CHECKER_CPP_CHECK,
+        GEN_CPP_CHECK,
+        FOLDER_CHECK,
+    )?;
+    let cases: usize = 10;
+
+    let mut cmd = Command::new(BINARY);
+    execute_command_check(
+        &mut cmd,
+        TARGET_FILE_CPP,
+        CHECKER_FILE_CPP,
+        GEN_FILE_CPP,
+        cases,
+    );
+
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("[OK]").count(cases));
+
+    Ok(())
+}
