+++ tests/cli/check_subcommand/mod.rs
@@ -1,12 +1,13 @@
 /*
  *  Quick Test: CLI for stress testing in competitive programming
- *  Copyright (C) 2021 - Luis Miguel Báez
+ *  Copyright (C) 2021-present / Luis Miguel Báez
  *  License: MIT (See the LICENSE file in the repository root directory)
  */
 
 pub mod codes;
+pub mod cpp;
 pub mod cpp_ce;
 pub mod cpp_mle;
-pub mod cpp_python;
 pub mod cpp_rte;
+pub mod python;
 pub mod python_rte;
