+++ tests/cli/cmp_subcommand/mod.rs
@@ -5,10 +5,11 @@
  */
 
 pub mod codes;
+pub mod cpp;
 pub mod cpp_ce;
 pub mod cpp_mle;
-pub mod cpp_python;
 pub mod cpp_rte;
+pub mod python;
 pub mod python_rte;
 // python throws an RTE (MemoryError) when it exceeds memory
 // pub mod python_mle;
