+++ tests/cli/stress_subcommand/mod.rs
@@ -5,8 +5,10 @@
  */
 
 pub mod codes;
+pub mod cpp;
 pub mod cpp_ce;
 pub mod cpp_mle;
 pub mod cpp_python;
 pub mod cpp_rte;
+pub mod python;
 pub mod python_rte;
